Refactor data_repository.py
- Move file loading from GitHub to `load.py`
- Move db saving/loading to `db.py`
