+++ pyard/data_repository.py
@@ -20,30 +20,30 @@
 #    > http://www.fsf.org/licensing/licenses/lgpl.html
 #    > http://www.opensource.org/licenses/lgpl-license.php
 #
-import sys
 import copy
-from collections import namedtuple
 import functools
 import sqlite3
-from urllib.error import HTTPError
+from collections import namedtuple
 
 import pandas as pd
 
+import pyard.load
+from pyard.smart_sort import smart_sort_comparator
 from . import db, broad_splits
-from .broad_splits import broad_splits_dna_mapping
-from .load import load_serology_broad_split_mapping
+from .load import (
+    load_g_group,
+    load_p_group,
+    load_allele_list,
+    load_serology_mappings,
+    load_latest_version,
+)
+from .misc import expression_chars
 from .misc import (
     get_2field_allele,
     get_3field_allele,
     number_of_fields,
     get_1field_allele,
 )
-from .misc import expression_chars, get_G_name, get_P_name
-
-# GitHub URL where IMGT HLA files are downloaded.
-from pyard.smart_sort import smart_sort_comparator
-
-IMGT_HLA_URL = "https://raw.githubusercontent.com/ANHIG/IMGTHLA/"
 
 ars_mapping_tables = [
     "dup_g",
@@ -84,44 +84,17 @@ def expression_reduce(df):
 
 def generate_ars_mapping(db_connection: sqlite3.Connection, imgt_version):
     if db.tables_exist(db_connection, ars_mapping_tables):
-        dup_g = db.load_dict(
-            db_connection, table_name="dup_g", columns=("allele", "g_group")
-        )
-        dup_lgx = db.load_dict(
-            db_connection, table_name="dup_lgx", columns=("allele", "lgx_group")
-        )
-        g_group = db.load_dict(
-            db_connection, table_name="g_group", columns=("allele", "g")
-        )
-        lgx_group = db.load_dict(
-            db_connection, table_name="lgx_group", columns=("allele", "lgx")
-        )
-        exon_group = db.load_dict(
-            db_connection, table_name="exon_group", columns=("allele", "exon")
-        )
-        p_not_g = db.load_dict(
-            db_connection, table_name="p_not_g", columns=("allele", "lgx")
-        )
-        return (
-            ARSMapping(
-                dup_g=dup_g,
-                dup_lgx=dup_lgx,
-                g_group=g_group,
-                lgx_group=lgx_group,
-                exon_group=exon_group,
-                p_not_g=p_not_g,
-            ),
-            None,
-        )
+        return db.load_ars_mappings(db_connection)
 
-    df = load_g_group(imgt_version)
-
+    df_g_group = load_g_group(imgt_version)
     df_p_group = load_p_group(imgt_version)
+
+    # Extract p group mapping
     p_group = df_p_group.set_index("A")["P"].to_dict()
 
-    # compare df_p_group["2d"] with df["2d"] to find 2-field alleles in the
+    # compare df_p_group["2d"] with df_g_group["2d"] to find 2-field alleles in the
     # P-group that aren't in the G-group
-    p_not_in_g = set(df_p_group["2d"]) - set(df["2d"])
+    p_not_in_g = set(df_p_group["2d"]) - set(df_g_group["2d"])
 
     # filter to find these 2-field alleles (2d) in the P-group data frame
     df_p_not_g = df_p_group[df_p_group["2d"].isin(p_not_in_g)]
@@ -133,7 +106,7 @@ def generate_ars_mapping(db_connection: sqlite3.Connection, imgt_version):
     # goal: identify 2-field alleles that are in multiple G-groups
 
     # group by 2d and G, and select the 2d column and count the columns
-    mg = df.drop_duplicates(["2d", "G"])["2d"].value_counts()
+    mg = df_g_group.drop_duplicates(["2d", "G"])["2d"].value_counts()
     # filter out the mg with count > 1, leaving only duplicates
     # take the index from the 2d version the data frame, make that a column
     # and turn that into a list
@@ -141,7 +114,7 @@ def generate_ars_mapping(db_connection: sqlite3.Connection, imgt_version):
 
     # Keep only the alleles that have more than 1 mapping
     dup_g = (
-        df[df["2d"].isin(multiple_g_list)][["G", "2d"]]
+        df_g_group[df_g_group["2d"].isin(multiple_g_list)][["G", "2d"]]
         .drop_duplicates()
         .groupby("2d", as_index=True)
         .agg("/".join)
@@ -149,12 +122,12 @@ def generate_ars_mapping(db_connection: sqlite3.Connection, imgt_version):
     )
 
     # multiple lgx
-    mlgx = df.drop_duplicates(["2d", "lgx"])["2d"].value_counts()
+    mlgx = df_g_group.drop_duplicates(["2d", "lgx"])["2d"].value_counts()
     multiple_lgx_list = mlgx[mlgx > 1].reset_index()["index"].to_list()
 
     # Keep only the alleles that have more than 1 mapping
     dup_lgx = (
-        df[df["2d"].isin(multiple_lgx_list)][["lgx", "2d"]]
+        df_g_group[df_g_group["2d"].isin(multiple_lgx_list)][["lgx", "2d"]]
         .drop_duplicates()
         .groupby("2d", as_index=True)
         .agg("/".join)
@@ -164,9 +137,9 @@ def generate_ars_mapping(db_connection: sqlite3.Connection, imgt_version):
     # Creating dictionaries with mac_code->ARS group mapping
     df_g = pd.concat(
         [
-            df[["2d", "G"]].rename(columns={"2d": "A"}),
-            df[["3d", "G"]].rename(columns={"3d": "A"}),
-            df[["A", "G"]],
+            df_g_group[["2d", "G"]].rename(columns={"2d": "A"}),
+            df_g_group[["3d", "G"]].rename(columns={"3d": "A"}),
+            df_g_group[["A", "G"]],
         ],
         ignore_index=True,
     )
@@ -174,9 +147,9 @@ def generate_ars_mapping(db_connection: sqlite3.Connection, imgt_version):
 
     df_lgx = pd.concat(
         [
-            df[["2d", "lgx"]].rename(columns={"2d": "A"}),
-            df[["3d", "lgx"]].rename(columns={"3d": "A"}),
-            df[["A", "lgx"]],
+            df_g_group[["2d", "lgx"]].rename(columns={"2d": "A"}),
+            df_g_group[["3d", "lgx"]].rename(columns={"3d": "A"}),
+            df_g_group[["A", "lgx"]],
         ]
     )
     lgx_group = df_lgx.set_index("A")["lgx"].to_dict()
@@ -184,174 +157,24 @@ def generate_ars_mapping(db_connection: sqlite3.Connection, imgt_version):
     # exon
     df_exon = pd.concat(
         [
-            df[["A", "3d"]].rename(columns={"3d": "exon"}),
+            df_g_group[["A", "3d"]].rename(columns={"3d": "exon"}),
         ]
     )
     exon_group = df_exon.set_index("A")["exon"].to_dict()
 
     # save
-    db.save_dict(
-        db_connection,
-        table_name="p_not_g",
-        dictionary=p_not_g,
-        columns=("allele", "lgx"),
+    return db.save_ars_mappings(
+        db_connection, dup_g, dup_lgx, exon_group, g_group, lgx_group, p_group, p_not_g
     )
-    db.save_dict(
-        db_connection,
-        table_name="dup_g",
-        dictionary=dup_g,
-        columns=("allele", "g_group"),
-    )
-    db.save_dict(
-        db_connection,
-        table_name="dup_lgx",
-        dictionary=dup_lgx,
-        columns=("allele", "lgx_group"),
-    )
-    db.save_dict(
-        db_connection, table_name="g_group", dictionary=g_group, columns=("allele", "g")
-    )
-    db.save_dict(
-        db_connection,
-        table_name="lgx_group",
-        dictionary=lgx_group,
-        columns=("allele", "lgx"),
-    )
-    db.save_dict(
-        db_connection,
-        table_name="exon_group",
-        dictionary=exon_group,
-        columns=("allele", "exon"),
-    )
-
-    return (
-        ARSMapping(
-            dup_g=dup_g,
-            dup_lgx=dup_lgx,
-            g_group=g_group,
-            lgx_group=lgx_group,
-            exon_group=exon_group,
-            p_not_g=p_not_g,
-        ),
-        p_group,
-    )
-
-
-def load_g_group(imgt_version):
-    # load the hla_nom_g.txt
-    ars_g_url = f"{IMGT_HLA_URL}{imgt_version}/wmda/hla_nom_g.txt"
-    df = pd.read_csv(ars_g_url, skiprows=6, names=["Locus", "A", "G"], sep=";").dropna()
-    # the G-group is named for its first allele
-    df["G"] = df["A"].apply(get_G_name)
-    # convert slash delimited string to a list
-    df["A"] = df["A"].apply(lambda a: a.split("/"))
-    # convert the list into separate rows for each element
-    df = df.explode("A")
-    #  A*   + 02:01   = A*02:01
-    df["A"] = df["Locus"] + df["A"]
-    df["G"] = df["Locus"] + df["G"]
-    # Create 2,3 field versions of the alleles
-    df["2d"] = df["A"].apply(get_2field_allele)
-    df["3d"] = df["A"].apply(get_3field_allele)
-    # lgx is 2 field version of the G group allele
-    df["lgx"] = df["G"].apply(get_2field_allele)
-
-    return df
-
-
-def load_p_group(imgt_version):
-    # load the hla_nom_p.txt
-    ars_p_url = f"{IMGT_HLA_URL}{imgt_version}/wmda/hla_nom_p.txt"
-    # example: C*;06:06:01:01/06:06:01:02/06:271;06:06P
-    df_p = pd.read_csv(
-        ars_p_url, skiprows=6, names=["Locus", "A", "P"], sep=";"
-    ).dropna()
-    # the P-group is named for its first allele
-    df_p["P"] = df_p["A"].apply(get_P_name)
-    # convert slash delimited string to a list
-    df_p["A"] = df_p["A"].apply(lambda a: a.split("/"))
-    df_p = df_p.explode("A")
-    # C* 06:06:01:01/06:06:01:02/06:271 06:06P
-    df_p["A"] = df_p["Locus"] + df_p["A"]
-    df_p["P"] = df_p["Locus"] + df_p["P"]
-    # C* 06:06:01:01 06:06P
-    # C* 06:06:01:02 06:06P
-    # C* 06:271 06:06P
-    df_p["2d"] = df_p["A"].apply(get_2field_allele)
-    # lgx has the P-group name without the P for comparison
-    df_p["lgx"] = df_p["P"].apply(get_2field_allele)
-    return df_p
 
 
 def generate_alleles_and_xx_codes_and_who(
     db_connection: sqlite3.Connection, imgt_version, ars_mappings, p_group
 ):
-    """
-    Checks to see if there's already an allele list file for the `imgt_version`
-    in the `data_dir` directory. If not, will download the file and create
-    a valid allele set and corresponding xx codes.
-
-    The format of the AlleleList file has a 6-line header with a header
-    on the 7th line
-    ```
-    # file: Allelelist.3290.txt
-    # date: 2017-07-10
-    # version: IPD-IMGT/HLA 3.29.0
-    # origin: https://github.com/ANHIG/IMGTHLA/Allelelist.3290.txt
-    # repository: https://raw.githubusercontent.com/ANHIG/IMGTHLA/Latest/allelelist/Allelelist.3290.txt
-    # author: WHO, Steven G. E. Marsh (steven.marsh@ucl.ac.uk)
-    AlleleID,Allele
-    HLA00001,A*01:01:01:01
-    HLA02169,A*01:01:01:02N
-    HLA14798,A*01:01:01:03
-    HLA15760,A*01:01:01:04
-    HLA16415,A*01:01:01:05
-    HLA16417,A*01:01:01:06
-    HLA16436,A*01:01:01:07
-    ```
-
-    :param db_connection: Database connection to the sqlite database
-    :param imgt_version: IMGT database version
-    :param ars_mappings: ARSMapping object to ARS mapping tables
-    :return: None, updates self
-    """
     if db.tables_exist(db_connection, code_mapping_tables):
-        valid_alleles = db.load_set(db_connection, "alleles")
+        return db.load_code_mappings(db_connection)
 
-        who_alleles = db.load_set(db_connection, "who_alleles")
-        who_group = db.load_dict(db_connection, "who_group", ("who", "allele_list"))
-        who_group = {k: v.split("/") for k, v in who_group.items()}
-
-        xx_codes = db.load_dict(db_connection, "xx_codes", ("allele_1d", "allele_list"))
-        xx_codes = {k: v.split("/") for k, v in xx_codes.items()}
-
-        exp_alleles = db.load_dict(
-            db_connection, "exp_alleles", ("exp_allele", "allele_list")
-        )
-        exp_alleles = {k: v.split("/") for k, v in exp_alleles.items()}
-
-        return valid_alleles, who_alleles, xx_codes, who_group, exp_alleles
-
-    # Create a Pandas DataFrame from the mac_code list file
-    # Skip the header (first 6 lines) and use only the Allele column
-    if imgt_version == "Latest":
-        allele_list_url = f"{IMGT_HLA_URL}Latest/Allelelist.txt"
-    else:
-        if imgt_version == "3130":
-            # 3130 was renamed to 3131 for Allelelist file only 🤷🏾‍
-            imgt_version = "3131"
-        allele_list_url = (
-            f"{IMGT_HLA_URL}Latest/allelelist/Allelelist.{imgt_version}.txt"
-        )
-
-    try:
-        allele_df = pd.read_csv(allele_list_url, header=6, usecols=["Allele"])
-    except HTTPError as e:
-        print(
-            f"Failed importing alleles for version {imgt_version} from {allele_list_url}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
+    allele_df = load_allele_list(imgt_version)
 
     # Create a set of valid alleles
     # All 2-field, 3-field and the original Alleles are considered valid alleles
@@ -364,12 +187,6 @@ def generate_alleles_and_xx_codes_and_who(
         )
     ]
     exp_alleles = exp_alleles.groupby("2d").apply(expression_reduce).dropna()
-
-    # flat_exp_alleles = {k: '/'.join(sorted(v, key=functools.cmp_to_key(smart_sort_comparator)))
-    #                 for k, v in exp_alleles.items()}
-    db.save_dict(
-        db_connection, "exp_alleles", exp_alleles, ("exp_allele", "allele_list")
-    )
 
     # Create valid set of alleles:
     # All full length alleles
@@ -397,19 +214,14 @@ def generate_alleles_and_xx_codes_and_who(
             else:
                 xx_codes[broad] = copy.deepcopy(xx_codes[split])
 
-    # Save this version of the valid alleles
-    db.save_set(db_connection, "alleles", valid_alleles, "allele")
     # Save this version of xx codes
     flat_xx_codes = {
         k: "/".join(sorted(v, key=functools.cmp_to_key(smart_sort_comparator)))
         for k, v in xx_codes.items()
     }
-    db.save_dict(db_connection, "xx_codes", flat_xx_codes, ("allele_1d", "allele_list"))
 
     # W H O
     who_alleles = allele_df["Allele"].to_list()
-    # Save this version of the WHO alleles
-    db.save_set(db_connection, "who_alleles", who_alleles, "allele")
 
     # Create WHO mapping from the unique alleles in the 1-field column
     allele_df["1d"] = allele_df["Allele"].apply(get_1field_allele)
@@ -441,8 +253,14 @@ def generate_alleles_and_xx_codes_and_who(
         k: "/".join(sorted(v, key=functools.cmp_to_key(smart_sort_comparator)))
         for k, v in who_group.items()
     }
-    db.save_dict(
-        db_connection, "who_group", flat_who_group, columns=("who", "allele_list")
+
+    db.save_code_mappings(
+        db_connection,
+        exp_alleles,
+        flat_who_group,
+        flat_xx_codes,
+        valid_alleles,
+        who_alleles,
     )
 
     return valid_alleles, who_alleles, xx_codes, who_group, exp_alleles
@@ -450,11 +268,7 @@ def generate_alleles_and_xx_codes_and_who(
 
 def generate_short_nulls(db_connection, who_group):
     if db.table_exists(db_connection, "shortnulls"):
-        shortnulls = db.load_dict(
-            db_connection, "shortnulls", ("shortnull", "allele_list")
-        )
-        shortnulls = {k: v.split("/") for k, v in shortnulls.items()}
-        return shortnulls
+        return db.load_shortnulls(db_connection)
 
     # shortnulls
     # scan WHO alleles for those with expression characters and make shortnull mappings
@@ -481,7 +295,8 @@ def generate_short_nulls(db_connection, who_group):
                 # e.g. DRB4*01:03N
                 shortnulls[a_shortnull] = "/".join(expression_alleles[a_shortnull])
 
-    db.save_dict(db_connection, "shortnulls", shortnulls, ("shortnull", "allele_list"))
+    db.save_shortnulls(db_connection, shortnulls)
+
     shortnulls = {k: v.split("/") for k, v in shortnulls.items()}
     return shortnulls
 
@@ -490,46 +305,6 @@ def generate_mac_codes(
     db_connection: sqlite3.Connection, refresh_mac: bool = False, load_mac: bool = True
 ):
     """
-    MAC files come in 2 different versions:
-
-    Martin: when they’re printed, the first is better for encoding and the
-    second is better for decoding. The entire list was maintained both as an
-    excel spreadsheet and also as a sybase database table. The excel was the
-    one that was printed and distributed.
-
-        **==> numer.v3.txt <==**
-
-        Sorted by the length and the the values in the list
-        ```
-        "LAST UPDATED: 09/30/20"
-        CODE	SUBTYPE
-
-        AB	01/02
-        AC	01/03
-        AD	01/04
-        AE	01/05
-        AG	01/06
-        AH	01/07
-        AJ	01/08
-        ```
-
-        **==> alpha.v3.txt <==**
-
-        Sorted by the code
-
-        ```
-        "LAST UPDATED: 10/01/20"
-        *	CODE	SUBTYPE
-
-            AA	01/02/03/05
-            AB	01/02
-            AC	01/03
-            AD	01/04
-            AE	01/05
-            AF	01/09
-            AG	01/06
-        ```
-
     :param db_connection: Database connection to the sqlite database
     :param refresh_mac: Refresh the database with newer MAC data ?
     :param load_mac: Should MAC be loaded at all
@@ -538,25 +313,10 @@ def generate_mac_codes(
     if load_mac:
         mac_table_name = "mac_codes"
         if refresh_mac or not db.table_exists(db_connection, mac_table_name):
-            # Load the MAC file to a DataFrame
-            mac_url = "https://hml.nmdp.org/mac/files/numer.v3.zip"
-            df_mac = pd.read_csv(
-                mac_url,
-                sep="\t",
-                compression="zip",
-                skiprows=3,
-                names=["Code", "Alleles"],
-                keep_default_na=False,
-            )
+            df_mac = pyard.load.load_mac_codes()
             # Create a dict from code to alleles
             mac = df_mac.set_index("Code")["Alleles"].to_dict()
-            # Save the mac dict to db
-            db.save_dict(
-                db_connection,
-                table_name=mac_table_name,
-                dictionary=mac,
-                columns=("code", "alleles"),
-            )
+            db.save_mac_codes(db_connection, mac, mac_table_name)
 
 
 def to_serological_name(locus_name: str):
@@ -580,28 +340,7 @@ def to_serological_name(locus_name: str):
 
 def generate_serology_mapping(db_connection: sqlite3.Connection, imgt_version):
     if not db.table_exists(db_connection, "serology_mapping"):
-        """
-        Read `rel_dna_ser.txt` file that contains alleles and their serological equivalents.
-
-        The fields of the Alleles->Serological mapping file are:
-           Locus - HLA Locus
-           Allele - HLA Allele Name
-           USA - Unambiguous Serological Antigen associated with allele
-           PSA - Possible Serological Antigen associated with allele
-           ASA - Assumed Serological Antigen associated with allele
-           EAE - Expert Assigned Exceptions in search determinants of some registries
-
-        EAE is ignored when generating the serology map.
-        """
-        rel_dna_ser_url = f"{IMGT_HLA_URL}{imgt_version}/wmda/rel_dna_ser.txt"
-        # Load WMDA serology mapping data from URL
-        df_sero = pd.read_csv(
-            rel_dna_ser_url,
-            sep=";",
-            skiprows=6,
-            names=["Locus", "Allele", "USA", "PSA", "ASA", "EAE"],
-            index_col=False,
-        )
+        df_sero = load_serology_mappings(imgt_version)
 
         # Remove 0 and ? from USA
         df_sero = df_sero[(df_sero["USA"] != "0") & (df_sero["USA"] != "?")]
@@ -658,39 +397,12 @@ def generate_serology_mapping(db_connection: sqlite3.Connection, imgt_version):
                 )
             )
 
-        # Save the serology mapping to db
-        db.save_dict(
-            db_connection,
-            table_name="serology_mapping",
-            dictionary=sero_mapping,
-            columns=("serology", "allele_list"),
-        )
+        db.save_serology_mappings(db_connection, sero_mapping)
 
 
 def generate_v2_to_v3_mapping(db_connection: sqlite3.Connection, imgt_version):
     if not db.table_exists(db_connection, "v2_mapping"):
-        # TODO: Create mapping table using both the allele list history and
-        #  deleted alleles as reference.
-        # Temporary Example
-        v2_to_v3_example = {
-            "A*0104": "A*01:04:01:01N",
-            "A*0105N": "A*01:04:01:01N",
-            "A*0111": "A*01:11N",
-            "A*01123": "A*01:123N",
-            "A*0115": "A*01:15N",
-            "A*0116": "A*01:16N",
-            "A*01160": "A*01:160N",
-            "A*01162": "A*01:162N",
-            "A*01178": "A*01:178N",
-            "A*01179": "A*01:179N",
-            "DRB5*02ZB": "DRB5*02:UTV",
-        }
-        db.save_dict(
-            db_connection,
-            table_name="v2_mapping",
-            dictionary=v2_to_v3_example,
-            columns=("v2", "v3"),
-        )
+        db.load_v2_v3_mappings(db_connection)
 
 
 def set_db_version(db_connection: sqlite3.Connection, imgt_version):
@@ -706,24 +418,14 @@ def set_db_version(db_connection: sqlite3.Connection, imgt_version):
     if version:
         return version
 
-    version = imgt_version
-
     if imgt_version == "Latest":
-        from urllib.request import urlopen
-
-        response = urlopen(
-            "https://raw.githubusercontent.com/ANHIG/IMGTHLA/Latest/release_version.txt"
-        )
-        for line in response:
-            l = line.decode("utf-8")
-            if l.find("version:") != -1:
-                # Version line looks like
-                # # version: IPD-IMGT/HLA 3.51.0
-                version = l.split()[-1].replace(".", "")
+        version = load_latest_version()
+    else:
+        version = imgt_version
 
     db.set_user_version(db_connection, int(version))
     print("Version:", version)
-    return db.get_user_version(db_connection)
+    return version
 
 
 def get_db_version(db_connection: sqlite3.Connection):
@@ -732,19 +434,8 @@ def get_db_version(db_connection: sqlite3.Connection):
 
 def generate_serology_broad_split_mapping(db_connection, imgt_version):
     if not db.table_exists(db_connection, "serology_broad_split_mapping"):
-        sero_mapping = load_serology_broad_split_mapping(imgt_version)
-        # Save the `splits` as a "/" delimited string to db
-        sero_splits = {sero: "/".join(splits) for sero, splits in sero_mapping.items()}
-        db.save_dict(
-            db_connection,
-            table_name="serology_broad_split_mapping",
-            dictionary=sero_splits,
-            columns=("serology", "splits"),
-        )
+        sero_mapping = pyard.load.load_serology_broad_split_mapping(imgt_version)
+        db.save_serology_broad_split_mappings(db_connection, sero_mapping)
         return sero_mapping
 
-    sero_mapping = db.load_dict(
-        db_connection, "serology_broad_split_mapping", ("serology", "splits")
-    )
-    sero_splits = {k: v.split("/") for k, v in sero_mapping.items()}
-    return sero_splits
+    return db.load_serology_broad_split_mappings(db_connection)
